// src/infrastructure/d-injection/types.ts
export enum InjectionValue {
  USER_REPOSITORY = 'UserRepository',
  PROFILE_REPOSITORY = 'ProfileRepository',
  ADDRESS_REPOSITORY = 'AddressRepository',
  SECURITY = 'SECURITY',
  COUNTRY_REPOSITORY = 'CountryRepository',
  CITY_REPOSITORY = 'CityRepository',
  BCRYPT = 'PasswordHasherService',
  CACHE = 'RedisRepository',
}